Move sortAlphabetically to public header
